Cast environment variable to integer.
Changed the 'shift_reservation_duration' configuration to cast the environment variable to an integer.

+++ config/cart-scheduler.php
@@ -2,7 +2,7 @@
 
 return [
     'version'                            => env('SELF_UPDATER_VERSION_INSTALLED', '0.0.1'),
-    'shift_reservation_duration'         => env('CA_SHIFT_RESERVATION_DURATION', 1),
+    'shift_reservation_duration'         => (int)env('CA_SHIFT_RESERVATION_DURATION', 1),
     'shift_reservation_duration_period'  => env('CA_SHIFT_RESERVATION_DURATION_PERIOD', 'MONTH'),
     'release_weekly_shifts_on_day'       => env('CA_SHIFT_RESERVATION_WEEK_RELEASE_DAY', 'SUN'),
     'release_new_shifts_at_time'         => env('CA_SHIFT_RESERVATION_NEW_RELEASE_TIME', '00:00:00'),
